Creating getter only in ExerciseSchema and setters for weights in every write model.

+++ src/main/java/springweb/trainingmanager/models/entities/Exercise.java
@@ -43,7 +43,7 @@ public class Exercise extends ExerciseSchema {
         this.repetition = repetition;
     }
 
-    public void setWeights(float weights){
+    public void setWeights(int weights){
         this.weights = weights;
     }
