+++ src/main/java/springweb/trainingmanager/models/viewmodels/exercise/ExerciseWrite.java
@@ -58,7 +58,7 @@ public class ExerciseWrite extends ExerciseSchema {
         this.repetition = repetition;
     }
 
-    public void setWeights(float weights){
+    public void setWeights(int weights){
         this.weights = weights;
     }
